Move more kernel API tests into kernel_tests
