+++ numba_dpex/tests/kernel_tests/test_private_memory.py
@@ -2,9 +2,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-import platform
 
-import dpctl
 import numpy as np
 import pytest
